Hopefully fixed api.yaml, converted int to uint64 for User and Photo IDs

+++ service/database/ban.go
@@ -6,7 +6,7 @@ import (
 
 var ErrAlreadyBanned = errors.New("the user is already banned")
 
-func (db *appdbimpl) Ban(userID int, bannedID int) error {
+func (db *appdbimpl) Ban(userID uint64, bannedID uint64) error {
 	sqlres, err := db.c.Exec("INSERT INTO Bans (userid, bannedid) VALUES (?, ?) ON CONFLICT DO NOTHING", userID, bannedID)
 	if err != nil {
 		return err
